+++ src/main/java/fact/features/HillasLength.java
@@ -22,7 +22,7 @@ public class HillasLength implements Processor {
 	
 		PixelDistribution2D dist = (PixelDistribution2D) input.get(distribution);
 
-		float width = (float) Math.sqrt(dist.getEigenVarianceY());
+		double width = (double) Math.sqrt(dist.getEigenVarianceY());
 	    input.put(outputKey, width);
 		
 	    return input;
